STRUCT: Move async method to client module
