+++ secedgar/client/_base.py
@@ -40,3 +40,8 @@ class AbstractClient(ABC):
     def batch_size(self):
         """Number of results per page searched. Increasing can improve speed for large requests."""
         pass  # pragma: no cover
+
+    @abstractmethod
+    async def wait_for_download_async(self, inputs):
+        """Asynchronously download links into files using rate limit."""
+        pass
